test(sentry): test_sentry_init.pyを責務別4ファイルに分割 (#532)
* test(sentry): test_sentry_init.pyを責務別4ファイルに分割

【背景】2922行に肥大化したtest_sentry_init.pyは責務が混在し可読性・保守性が低下していた。

【変更内容】init/primitives(機密キー判定)/values(値scrub)/events(イベント整形)の責務軸で4ファイルに分割。あわせてIssue #531の対象3クラス(単一テスト・無状態: TestScrubbedEventFieldsCompleteness/TestScrubUrlIpv6/TestBeforeSendLoggerEventId)をmodule関数へ変換。

【影響範囲】テストファイルのみの構造変更で本番コード(utils/)は無変更。テスト件数・カバレッジは変換前後で不変。

品質ゲート: 1443 tests passed / 97.58% coverage / ruff 0 errors / mypy 0 errors

Closes #531 (Issue)

* test(sentry): レビュー対応でテスト配置とdocstringを修正 (PR #532 review)

【背景】PR #532 のCodeRabbitレビューで、責務分割方針とsentinelテストの
docstring正確性に関する2件の指摘を受けた。

【変更内容】_before_send のイベント整形を検証するテストを
test_sentry_scrub_primitives.py から責務が一致する
test_sentry_scrub_events.py へ移動し、未使用となったimportを整理。
併せてsentinelテストのdocstringにあった編集残骸（文の途中切れ）を修復。

品質ゲート: 1443 tests passed / 97.58% coverage / ruff 0 errors / mypy 0 errors

Refs #532 (PR)

* test(sentry): 失敗メッセージのログダンプを削除 (PR #532 review)

【背景】PR#532 のレビュー指摘(Security & Privacy / Major)。assertion
失敗時に mock_logger.error.call_args_list 全体をダンプすると、将来の実装
変更で URL・例外・機密値が CI 失敗ログへ漏洩しうる。本テストは Sentry の
PII スクラブ抑止スイートであり、漏洩防止が最優先の箇所。

【変更内容】失敗メッセージからログ呼び出し全体のダンプを削除し、安全な
件数(len(drop_event_calls))のみ表示。assert 検証ロジックは不変。

品質ゲート: 1443 tests passed / 97.58% coverage / ruff 0 errors / mypy 0 errors

Refs #532 (PR)

* test(sentry): token判定コメントを実測に訂正 (PR #532 review)

【背景】_SENSITIVE_KEY_PATTERN.search("token") は直接マッチする
（prefix `^` と suffix `$` がともにゼロ幅アンカーで境界を満たす）にも
かかわらず、コメント・テスト名・docstring は「パターン非一致で
compact fallback 経由」と誤説明していた。

【変更内容】3箇所を実測事実に訂正。テスト名を matched_by_pattern へ
改名し、陳腐化した行番号参照 (sentry_init.py:210-212) を削除。
アサーション・テストロジックは不変。

品質ゲート: 1443 tests passed / 97.58% coverage / ruff 0 errors / mypy 0 errors

Refs #532 (PR)

---------
